src/main/java/io/ljunggren/unitConversion/temperature/unit/si/Celsius.java
package io.ljunggren.unitConversion.temperature.unit.si;

import java.util.function.Function;

import io.ljunggren.unitConversion.System;
import io.ljunggren.unitConversion.temperature.Temperature;

public class Celsius implements Temperature {

    @Override
    public Function<Double, Double> toBaseUnit() {
        return x -> x;
    }

    @Override
    public Function<Double, Double> fromBaseUnit() {
        return x -> x;
    }

    @Override
    public System system() {
        return System.SI;
    }

    @Override
    public String label() {
        return "C";
    }

}